1-producer-service/requirements.txt
requests
kafka-python
yfinance
pandas
tqdm
schedule
requests